block_dev: Use 64-bit value for block device size
Also replace block_dev_ioctl() for getting device size and block size
with  simple  handlers  (ioctl() is  not suitable there anyway: those
functions operate with block devices, not file descriptors or pathes)

+++ src/fs/driver/iso9660/iso9660.h
@@ -59,7 +59,7 @@ typedef int64_t off64_t;
 
 typedef struct cdfs {
 	void *bdev;
-	int blks;
+	uint64_t blks;
 	int volblks;
 	int vdblk;
 	int joliet;
